cord tests: Suppress GCC warnings in 'de' app
* cord/tests/de.c (retrieve_line): Change type of "eol" local variable
from int to size_t (to prevent "comparison between signed and unsigned"
compiler warning).
* cord/tests/de_win.c (WinMain): Cast GetLastError() result to
unsigned int (to match printf format specifier).
* cord/tests/de_win.c (AboutBoxCallback): Suppress "unused parameter"
compiler warning for "lParam" (by casting the later to void).

+++ cord/tests/de.c
@@ -242,7 +242,7 @@ CORD retrieve_line(CORD s, size_t pos, unsigned column)
 {
     CORD candidate = CORD_substr(s, pos, column + COLS);
                         /* avoids scanning very long lines      */
-    int eol = CORD_chr(candidate, 0, '\n');
+    size_t eol = CORD_chr(candidate, 0, '\n');
     int len;
 
     if (eol == CORD_NOT_FOUND) eol = CORD_len(candidate);
